requirements.txt
Flask == 2.2.3
Werkzeug == 2.2.3
Pillow == 9.5.0
